new sidepanel + refactor (wip)
